requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.3
django-countries==7.6.1
django-restframework==0.0.1
idna==3.10
pillow==11.0.0
psycopg2==2.9.10
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.2
stripe==11.4.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
